api with bugs LOL

+++ src/main/java/eat_it_server/repository/BracketRepository.java
@@ -3,5 +3,5 @@ package eat_it_server.repository;
 import eat_it_server.model.Bracket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BracketRepository extends JpaRepository<Bracket, Long> {
+public interface BracketRepository extends JpaRepository<Bracket, Integer> {
 }
